requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cffi==1.16.0
discord.py==2.4.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
h3==3.7.7
idna==3.7
multidict==6.0.5
numpy==2.0.0
pycparser==2.22
python-weather==2.0.3
pytz==2024.1
setuptools==70.1.0
timezonefinder==6.5.2
yarl==1.9.4
